requirements.txt
langchain
faiss-cpu
pypdf
tiktoken
transformers
databutton
streamlit
requests